domain, persistence and controllers crud

+++ soccerleague/domain/src/main/java/br/com/projects/domain/business/publico/match/spi/CrudMatch.java
@@ -5,7 +5,7 @@ import br.com.projects.domain.business.publico.match.DMatch;
 
 import java.util.Collection;
 
-public interface CrudMatch extends SimpleCrud<DMatch, Long> {
+public interface CrudMatch extends SimpleCrud<DMatch, Integer> {
 
     Collection<? extends DMatch> findByHomeTeamAndAwayTeamAndMatchDay (Integer homeTeamId, Integer awayTeamId, Integer matchDayId);
 }
